+++ components/collector/src/source_collectors/gatling/base.py
@@ -7,24 +7,30 @@ from collector_utilities.type import URL
 from model import SourceResponses
 
 
-class GatlingJSONCollector(JSONFileSourceCollector):
+class GatlingCollector(SourceCollector):
+    """Base class for Gatling collectors."""
+
+    INDEX_HTML = "/index.html"
+    FILEPATH = "subclass responsibility"
+
+    async def _api_url(self) -> URL:
+        """Extend to translate the HTML URL into a URL to the Gatling file that the collector needs."""
+        api_url = await super()._api_url()
+        if api_url.endswith(self.INDEX_HTML):
+            return URL(api_url.removesuffix(self.INDEX_HTML) + self.FILEPATH)
+        return api_url
+
+
+class GatlingJSONCollector(JSONFileSourceCollector, GatlingCollector):
     """Base class for Gatling collectors that read the stats.json file."""
 
-    async def _api_url(self) -> URL:
-        """Extend to translate the HTML URL into a URL to the stats.json file."""
-        api_url = await super()._api_url()
-        api_url.removesuffix("/index.html")
-        return URL(api_url + "/js/stats.json")
+    FILEPATH = "/js/stats.json"
 
 
-class GatlingLogCollector(SourceCollector):
+class GatlingLogCollector(GatlingCollector):
     """Base class for Gatling collectors that read the simulation.log file."""
 
-    async def _api_url(self) -> URL:
-        """Extend to translate the HTML URL into a URL to the simulation.log file."""
-        api_url = await super()._api_url()
-        api_url.removesuffix("/index.html")
-        return URL(api_url + "/simulation.log")
+    FILEPATH = "/simulation.log"
 
     @classmethod
     async def _timestamps(cls, responses: SourceResponses) -> set[int]: